+++ lib/ml_dataframe.dart
@@ -1,3 +1,4 @@
 export 'package:ml_dataframe/src/data_frame/data_frame.dart';
 export 'package:ml_dataframe/src/data_frame/factories/from_csv.dart';
+export 'package:ml_dataframe/src/data_frame/factories/from_json.dart';
 export 'package:ml_dataframe/src/data_frame/series.dart';